[AF013] Added RelativeLayout with three Textviews
In activity_main.xml the main Layout is now a Relative Layout. The three textviews are arranged with the layout_below = @id attribute below each other.

The text content is safed int the strings.xml and size attributes like padding or lineSpacingExtra are extracted to a dimens.xml file. On this way the attributes can be easier reused.