[AF0103] LinearLayout with views in ScrollIView
A Scroll View can only "posess" one view.

To add multiple Views to a ScrollView, they have to be wrapped up in a common  Layout.

I used the LinearLayout to hold the subtitle#, the article and a button.
A LinearLayout need the attribute orientation to be set.
Then the Linear Layout is passed to the ScrollView.

The TextView @id/tV_article has now the attribute android:autoLink ="web" to identify a web link inside it content. The link can then by used as hyperlink and opens the web content of it.

The coding challenge is finished.